Initial working version of webrtc helper
Provides a translation interface between the webrtc/libjingle
peerconnection_client (which the new fbstream is based on) and the
SVMP protobuf wire format.

On one side it emulates peerconnection_server's HTTP/JSON based interface.
On the other it connects using the SVMP protocol to an upstream server,
such as the SVMP input event server.

Change-Id: I1f61236ef60f5dcc631bb05008918de5a857154e