Add Android.mk and launch script for in-tree build
Android.mk uses the local copy of netty instead of the one that the
input server provides. The Netty 4 API hadn't stabilized for the pre-
release version of the lib that the input server includes. We can't
update that copy to final though since the inputer server and the two
pre-existing handlers use classes that were removed in 4 final.

Change-Id: I6b3292c32705f258dae3c72cff23263783260683